+++ R/run_checks.R
@@ -204,8 +204,16 @@ run_congruence_checks <- function(directory = here::here(),
            warning = function(w) {
              warn_count <<- warn_count + 1
              cli::cli_bullets(c(w$message, w$body))
-           }
-  )
+           })
+  tryCatch(test_datatable_url_attributes(metadata),
+           error = function(e) {
+             err_count <<-err_count + 1
+             cli::cli_bullets(c(e$message, e$body))
+           },
+           warning = function(w) {
+             warn_count <<- warn_count + 1
+             cli::cli_bullets(c(w$message, w$body))
+           })
   tryCatch(test_publisher(metadata),
            error = function(e) {
              err_count <<- err_count + 1